+++ clang-tools-extra/clangd/TidyProvider.cpp
@@ -221,6 +221,7 @@ TidyProvider disableUnusableChecks(llvm::ArrayRef<std::string> ExtraBadChecks) {
       // Check uses dataflow analysis, which might hang/crash unexpectedly on
       // incomplete code.
       "-bugprone-unchecked-optional-access",
+      "-bugprone-null-check-after-dereference",
       "-abseil-unchecked-statusor-access");
 
   size_t Size = BadChecks.size();